Update series ACL in index directly
